# apps/go-api/internal/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gravity Room API",
    "description": "REST API for the Gravity Room strength training tracker.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        }
      },
...
      "ProgramInstanceResponse": {
        "type": "object",
        "required": ["id", "programId", "name", "config", "metadata", "status", "results", "undoHistory", "resultTimestamps", "completedDates", "definitionId", "customDefinition", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "programId": { "type": "string" },
          "name": { "type": "string" },
          "config": { "type": "object", "description": "Program configuration (JSONB)" },
          "metadata": { "description": "Arbitrary metadata (JSONB)", "nullable": true },
          "status": { "type": "string", "enum": ["active", "completed", "archived"] },
          "results": { "type": "object", "additionalProperties": true },
          "undoHistory": { "type": "array", "items": { "type": "object" } },
          "resultTimestamps": { "type": "object", "additionalProperties": { "type": "string" } },
          "completedDates": { "type": "object", "additionalProperties": { "type": "string" } },
          "definitionId": { "type": "string", "nullable": true },
          "customDefinition": { "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ProgramDefinitionResponse": {
        "type": "object",
        "required": ["id", "userId", "definition", "status", "createdAt", "updatedAt", "deletedAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "definition": { "type": "object", "description": "Program definition (JSONB)" },
          "status": { "type": "string", "enum": ["draft", "pending_review", "approved", "rejected"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ProgramDefinitionListResponse": {
        "type": "object",
        "required": ["data", "total"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } },
          "total": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/programs/{id}/results": {
      "post": {
        "tags": ["Results"],
        "summary": "Record workout result",
        "description": "Record a result for a specific workout slot. Rate: 60/min per userId.",
        "operationId": "recordResult",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["workoutIndex", "slotId", "result"],
                "properties": {
                  "workoutIndex": { "type": "integer", "minimum": 0 },
                  "slotId": { "type": "string" },
                  "result": { "type": "string" },
                  "amrapReps": { "type": "integer", "nullable": true },
                  "rpe": { "type": "integer", "nullable": true },
                  "setLogs": { "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Result recorded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramInstanceResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Program not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/programs/{id}/results/{workoutIndex}/{slotId}": {
      "delete": {
        "tags": ["Results"],
        "summary": "Delete result",
        "description": "Delete a specific workout result. Rate: 20/min per userId.",
        "operationId": "deleteResult",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "workoutIndex", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "slotId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Result deleted" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/program-definitions": {
      "get": {
        "tags": ["Program Definitions"],
        "summary": "List definitions",
        "description": "List user's program definitions with offset pagination. Rate: 100/min per userId.",
        "operationId": "listDefinitions",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": { "description": "Definition list", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionListResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["Program Definitions"],
        "summary": "Create definition",
        "description": "Create a new program definition. Rate: 5/hour per userId. Max 10 per user.",
        "operationId": "createDefinition",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["definition"],
                "properties": { "definition": { "type": "object" } }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Definition created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/program-definitions/fork": {
      "post": {
        "tags": ["Program Definitions"],
        "summary": "Fork definition",
        "description": "Fork an existing definition (catalog or user-owned). Rate: 10/hour per userId.",
        "operationId": "forkDefinition",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sourceId", "sourceType"],
                "properties": {
                  "sourceId": { "type": "string" },
                  "sourceType": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Definition forked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Source not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Limit exceeded", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/program-definitions/{id}": {
      "get": {
        "tags": ["Program Definitions"],
        "summary": "Get definition",
        "description": "Get a specific program definition. Rate: 100/min per userId.",
        "operationId": "getDefinition",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Definition detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Program Definitions"],
        "summary": "Update definition",
        "description": "Replace definition content. Rate: 20/hour per userId.",
        "operationId": "updateDefinition",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["definition"],
                "properties": { "definition": { "type": "object" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated definition", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Program Definitions"],
        "summary": "Delete definition",
        "description": "Delete a program definition. Fails if active instances exist. Rate: 20/hour per userId.",
        "operationId": "deleteDefinition",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Active instances exist", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/program-definitions/{id}/status": {
      "patch": {
        "tags": ["Program Definitions"],
        "summary": "Update definition status",
        "description": "Transition definition status (draft/pending_review/approved/rejected). Admin users have elevated transitions. Rate: 20/hour per userId.",
        "operationId": "updateDefinitionStatus",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { "type": "string", "enum": ["draft", "pending_review", "approved", "rejected"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProgramDefinitionResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Invalid transition", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}